[FIX] Upgrade mui material try to fix pb with jest
