+++ src/__tests__/aqavit.test.tsx
@@ -1,18 +1,12 @@
 import React from 'react';
-import { render, waitFor } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest'
 import { axe } from 'vitest-axe';
-import Index, { Head } from '../index';
+import AQAvit, { Head } from '../pages/aqavit';
 
-const mockLatestLTS = {
-  mostRecentLts: {
-    version: 1
-  }
-};
-
-describe('Index page', () => {
+describe('AQAvit page', () => {
   it('renders correctly', () => {
-    const { container } = render(<Index data={mockLatestLTS} />);
+    const { container } = render(<AQAvit />);
     // eslint-disable-next-line
     const pageContent = container.querySelector('main');
 
@@ -23,11 +17,11 @@ describe('Index page', () => {
     const { container } = render(<Head />);
     // eslint-disable-next-line
     const title = container.querySelector('title');
-    expect(title?.textContent).toEqual('Home | Adoptium');
+    expect(title?.textContent).toEqual('Eclipse AQAvit | Adoptium');
   });
 
   it('has no accessibility violations', async () => {
-    const { container } = render(<Index data={mockLatestLTS} />);
+    const { container } = render(<AQAvit />);
     const results = await axe(container);
     expect(results).toHaveNoViolations();
   });
