+++ src/__tests__/members.test.tsx
@@ -1,12 +1,29 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import { describe, expect, it } from 'vitest'
+import { describe, expect, it, vi } from 'vitest'
 import { axe } from 'vitest-axe';
-import Join, { Head } from '../join';
+import Members, { Head } from '../pages/members';
 
-describe('Join page', () => {
+vi.mock('../util/shuffle', () => {
+  return {
+    shuffle: (array) => {
+      array = [
+        {
+          name: 'mock_member',
+          logo: 'mock_logo.png',
+          url: 'https://mock.com',
+          tier: 'mock_tier',
+        }
+      ]
+      return array
+    }
+  };
+});
+
+describe('Members page', () => {
   it('renders correctly', () => {
-    const { container } = render(<Join />);
+    const { container } = render(<Members />);
+
     // eslint-disable-next-line
     const pageContent = container.querySelector('main');
 
@@ -17,11 +34,11 @@ describe('Join page', () => {
     const { container } = render(<Head />);
     // eslint-disable-next-line
     const title = container.querySelector('title');
-    expect(title?.textContent).toEqual('Join | Adoptium');
+    expect(title?.textContent).toEqual('Adoptium Working Group Members | Adoptium');
   });
 
   it('has no accessibility violations', async () => {
-    const { container } = render(<Join />);
+    const { container } = render(<Members />);
     const results = await axe(container);
     expect(results).toHaveNoViolations();
   });
